refactor: move curve-common features into curve-ui-kit
