+++ apps/dao/src/layout/Header.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom'
 import { CONNECT_STAGE } from '@/constants'
 import { getLocaleFromUrl, getNetworkFromUrl, getRestFullPathname } from '@/utils/utilsRouter'
 import { _parseRouteAndIsActive, isLoading } from '@/ui/utils'
-import { getWalletSignerAddress, useConnectWallet } from '@/common/features/connect-wallet'
+import { getWalletSignerAddress, useConnectWallet } from '@ui-kit/features/connect-wallet'
 import networks, { visibleNetworksList } from '@/networks'
 import useLayoutHeight from '@/hooks/useLayoutHeight'
 import useStore from '@/store/useStore'
