+++ apps/lend/src/layout/Header.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom'
 import { CONNECT_STAGE } from '@/constants'
 import { getParamsFromUrl, getRestFullPathname, getRestPartialPathname } from '@/utils/utilsRouter'
 import { _parseRouteAndIsActive, FORMAT_OPTIONS, formatNumber, isLoading } from '@/ui/utils'
-import { getWalletSignerAddress, useConnectWallet } from '@/common/features/connect-wallet'
+import { getWalletSignerAddress, useConnectWallet } from '@ui-kit/features/connect-wallet'
 import networks, { visibleNetworksList } from '@/networks'
 import useStore from '@/store/useStore'
 import { useTvl } from '@/entities/chain'
