+++ apps/lend/src/layout/index.tsx
@@ -5,7 +5,7 @@ import { CONNECT_STAGE, ROUTE } from '@/constants'
 import { layoutHeightKeys } from '@/store/createLayoutSlice'
 import { getNetworkFromUrl } from '@/utils/utilsRouter'
 import { isFailure, isLoading } from '@/ui/utils'
-import { getWalletChainId, useConnectWallet } from '@/common/features/connect-wallet'
+import { getWalletChainId, useConnectWallet } from '@ui-kit/features/connect-wallet'
 import useStore from '@/store/useStore'
 import Header from '@/layout/Header'
 import { Footer } from 'curve-ui-kit/src/widgets/Footer'
