+++ apps/lend/src/networks.ts
@@ -1,6 +1,6 @@
 import sortBy from 'lodash/sortBy'
 import { getBaseNetworksConfig, NETWORK_BASE_CONFIG } from '@/ui/utils'
-import { ChainOption } from '@/common/features/switch-chain'
+import { ChainOption } from '@ui-kit/features/switch-chain'
 
 const DEFAULT_NETWORK_CONFIG = {
   hideMarketsInUI: {},
