+++ apps/lend/src/pages/_app.tsx
@@ -8,7 +8,7 @@ import 'focus-visible'
 import '@/globals.css'
 import { HashRouter } from 'react-router-dom'
 import type { AppProps } from 'next/app'
-import { connectWalletLocales, initOnboard } from '@/common/features/connect-wallet'
+import { connectWalletLocales, initOnboard } from '@ui-kit/features/connect-wallet'
 import { persister, queryClient } from '@ui-kit/lib/api/query-client'
 import { ThemeProvider } from 'curve-ui-kit/src/shared/ui/ThemeProvider'
 import GlobalStyle from '@/globalStyle'
