+++ apps/loan/src/networks.ts
@@ -1,7 +1,7 @@
 import sortBy from 'lodash/sortBy'
 import { getBaseNetworksConfig, NETWORK_BASE_CONFIG } from '@/ui/utils'
 import curvejsApi from '@/lib/apiCrvusd'
-import { ChainOption } from '@/common/features/switch-chain'
+import { ChainOption } from '@ui-kit/features/switch-chain'
 
 const DEFAULT_NETWORK_CONFIG = {
   api: curvejsApi,
