+++ apps/main/src/layout/default/Header.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom'
 import { CONNECT_STAGE, ROUTE } from '@/constants'
 import { _parseRouteAndIsActive, FORMAT_OPTIONS, formatNumber, isLoading } from '@/ui/utils'
 import { useParamsFromUrl, useRestPartialPathname } from '@/utils/utilsRouter'
-import { getWalletSignerAddress, useConnectWallet } from '@/common/features/connect-wallet'
+import { getWalletSignerAddress, useConnectWallet } from '@ui-kit/features/connect-wallet'
 import useStore from '@/store/useStore'
 import { Header as NewHeader } from '@/common/widgets/Header'
 import { NavigationSection } from '@/common/widgets/Header/types'
