+++ packages/curve-common/src/widgets/Header/DesktopHeader.tsx
@@ -1,14 +1,14 @@
 import { AppBar, Toolbar } from '@mui/material'
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
-import { ConnectWalletIndicator } from '../../features/connect-wallet'
-import { ChainSwitcher } from '../../features/switch-chain'
+import { ConnectWalletIndicator } from 'curve-ui-kit/src/features/connect-wallet'
+import { ChainSwitcher } from 'curve-ui-kit/src/features/switch-chain'
 import { AppButtonLinks } from './AppButtonLinks'
 import { HeaderLogo } from './HeaderLogo'
 import { HeaderStats } from './HeaderStats'
 import { PageTabs } from './PageTabs'
-import { ThemeSwitcherButton } from '../../features/switch-theme'
-import { AdvancedModeSwitcher } from '../../features/switch-advanced-mode'
+import { ThemeSwitcherButton } from 'curve-ui-kit/src/features/switch-theme'
+import { AdvancedModeSwitcher } from 'curve-ui-kit/src/features/switch-advanced-mode'
 import { BaseHeaderProps } from './types'
 import { DEFAULT_BAR_SIZE } from 'curve-ui-kit/src/themes/components'
 import { useState } from 'react'
