+++ packages/curve-common/src/widgets/Header/MobileTopBar.tsx
@@ -1,8 +1,7 @@
-import { ChainSwitcher, ChainSwitcherProps } from '../../features/switch-chain'
+import { ChainSwitcher, ChainSwitcherProps } from 'curve-ui-kit/src/features/switch-chain'
 import Box from '@mui/material/Box'
 import { MenuToggleButton } from './MenuToggleButton'
 import { HeaderLogo } from './HeaderLogo'
-import React from 'react'
 import type { Theme, SxProps } from '@mui/system'
 import { AppName } from 'curve-ui-kit/src/shared/routes'
