+++ packages/curve-common/src/widgets/Header/SideBarFooter.tsx
@@ -1,14 +1,14 @@
 import { BaseHeaderProps } from './types'
 import Box, { type BoxProps } from '@mui/material/Box'
-import { ThemeSwitcherButtons } from '../../features/switch-theme'
-import { ConnectWalletIndicator } from '../../features/connect-wallet'
+import { ThemeSwitcherButtons } from 'curve-ui-kit/src/features/switch-theme'
+import { ConnectWalletIndicator } from 'curve-ui-kit/src/features/connect-wallet'
 import type { SxProps, Theme } from '@mui/system'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import Accordion from '@mui/material/Accordion'
 import AccordionDetails from '@mui/material/AccordionDetails'
 import AccordionSummary from '@mui/material/AccordionSummary'
 import Typography from '@mui/material/Typography'
-import { AdvancedModeSwitcher } from '../../features/switch-advanced-mode'
+import { AdvancedModeSwitcher } from 'curve-ui-kit/src/features/switch-advanced-mode'
 import { t } from '@lingui/macro'
 import { GearIcon } from 'curve-ui-kit/src/shared/icons/GearIcon'
