+++ packages/curve-common/tsconfig.json
@@ -10,7 +10,6 @@
     "noEmit": true,
     "strict": true,
     "paths": {
-      "@ui-kit/*": ["../../../packages/curve-ui-kit/src/*"],
       "@/common/*": ["../../../packages/curve-common/src/*"]
     }
   },
