+++ packages/curve-ui-kit/package.json
@@ -17,10 +17,27 @@
     "@mui/icons-material": "^6.1.4",
     "@mui/material": "^6.1.4",
     "@mui/utils": "^6.1.4",
+    "@safe-global/safe-apps-provider": "^0.18.3",
+    "@safe-global/safe-apps-sdk": "^9.1.0",
     "@tanstack/query-sync-storage-persister": "^5.59.0",
     "@tanstack/react-query": "^5.59.0",
     "@tanstack/react-query-devtools": "^5.59.0",
     "@tanstack/react-query-persist-client": "^5.59.0",
+    "@web3-onboard/bitget": "^2.1.1",
+    "@web3-onboard/coinbase": "^2.4.1",
+    "@web3-onboard/core": "^2.22.3",
+    "@web3-onboard/fortmatic": "^2.1.1",
+    "@web3-onboard/frontier": "^2.1.1",
+    "@web3-onboard/gnosis": "^2.3.2",
+    "@web3-onboard/injected-wallets": "^2.11.2",
+    "@web3-onboard/ledger": "^2.7.1",
+    "@web3-onboard/metamask": "^2.1.3",
+    "@web3-onboard/phantom": "^2.1.1",
+    "@web3-onboard/react": "^2.9.3",
+    "@web3-onboard/torus": "^2.3.1",
+    "@web3-onboard/trezor": "^2.4.6",
+    "@web3-onboard/trust": "^2.1.2",
+    "@web3-onboard/walletconnect": "^2.6.1",
     "ethers": "^6.11.0",
     "react": "*",
     "react-dom": "*"
@@ -30,6 +47,8 @@
     "react-dom": "*"
   },
   "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@babel/runtime": "^7.25.6",
     "@chromatic-com/storybook": "^2.0.2",
     "@storybook/addon-a11y": "^8.3.5",
     "@storybook/addon-docs": "^8.3.5",
