+++ packages/curve-ui-kit/src/features/connect-wallet/lib/locales/zh-Hans.ts
