+++ packages/curve-ui-kit/src/features/connect-wallet/ui/ConnectWalletIndicator.tsx
