+++ packages/curve-ui-kit/src/features/switch-chain/ui/ChainIcon.tsx
