+++ packages/curve-ui-kit/src/features/switch-chain/ui/ChainSwitcher.tsx
@@ -2,15 +2,15 @@ import IconButton from '@mui/material/IconButton'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
 import { useEffect, useMemo } from 'react'
-import { ThemeKey } from 'curve-ui-kit/src/themes/basic-theme'
-import { useSwitch } from 'curve-ui-kit/src/hooks/useSwitch'
-import { ModalDialog } from 'curve-ui-kit/src/shared/ui/ModalDialog'
+import { ThemeKey } from '@ui-kit/themes/basic-theme'
+import { useSwitch } from '@ui-kit/hooks/useSwitch'
+import { ModalDialog } from '@ui-kit/shared/ui/ModalDialog'
 import { t } from '@lingui/macro'
 import { SettingsButton } from './SettingsButton'
 import { ChainIcon } from './ChainIcon'
 import { ChainList } from './ChainList'
 import { ChainSettings } from './ChainSettings'
-import { useLocalStorage } from 'curve-ui-kit/src/hooks/useLocalStorage'
+import { useLocalStorage } from '@ui-kit/hooks/useLocalStorage'
 
 export type ChainOption<TChainId> = {
   chainId: TChainId
